project/recruiters/backend/services/applications-service/requirements.txt
fastapi[all]==0.115.9
sqlalchemy==2.0.38
asyncpg==0.27.0
gunicorn==23.0.0
opentelemetry-instrumentation-fastapi==0.48b0
pytest==7.0.1
testcontainers==4.9.1
pydantic==2.8.2
python-dotenv==1.0.1
psycopg2-binary==2.9.3